Bomb bay doors with strings
I didn't have the exact same first distance as Cody did because I didn't have the file when I started making these files. This system should be be easily modified. If new parts need  to be make, the main section will not need to be reprinted. For printing, there is a file in Millimeters and Inches for any type of 3D printer. Also, the short brace file should be printed twice for it to work. The spool system was made larger than it will be in the final design so only the spool and its cover needs to be reprinted if a different size spool needs to be tested.